services/routing-service/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
redis==5.0.1
python-dotenv==1.0.0
httpx==0.25.1
googlemaps==4.10.0
